bank-balance-with-exactly-once/src/test/java/com/udemy/kafka/streams/course/BankTransactionsProducerTests.java
package com.udemy.kafka.streams.course;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.JsonNode;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.apache.kafka.clients.producer.ProducerRecord;
import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

public class BankTransactionsProducerTests {
    @Test
    public void newRandomTransactionsTests() {
        ProducerRecord<String, String> record = BankTransactionsProducer.newRandomTransaction("john");
        String key = record.key();
        String value = record.value();

        assertEquals(key, "john");

        ObjectMapper mapper = new ObjectMapper();
        try {
            JsonNode node = mapper.readTree(value);
            assertEquals(node.get("name").asText(), "john");
            assertTrue(node.get("amount").asInt() < 100, "Amount should be less than 100");
        } catch (JsonProcessingException e) {
            throw new RuntimeException(e);
        }
        System.out.println(value);
    }
}
